fix(positions): Move cache registration logic to position service
