+++ src/cache/cache.module.ts
@@ -7,5 +7,6 @@ import { CacheService } from './cache.service';
 @Module({
   imports: [DiscoveryModule, NestCacheModule.register({ ttl: 0, max: Number.MAX_SAFE_INTEGER, isGlobal: true })],
   providers: [CacheService, CacheOnIntervalService],
+  exports: [CacheService, CacheOnIntervalService],
 })
 export class CacheModule {}
