+++ src/position/position-fetcher.decorator.ts
@@ -1,11 +1,8 @@
 import { applyDecorators, Injectable, SetMetadata } from '@nestjs/common';
 
-import { CacheOnIntervalBuilder } from '~cache/cache-on-interval.decorator';
 import { Network } from '~types/network.interface';
 
 import { ContractType } from './contract.interface';
-import { PositionFetcher as IPositionFetcher } from './position-fetcher.interface';
-import { Position } from './position.interface';
 
 export const POSITION_FETCHER_APP = 'POSITION_FETCHER_APP';
 export const POSITION_FETCHER_GROUP = 'POSITION_FETCHER_GROUP';
@@ -16,13 +13,6 @@ export const POSITION_FETCHER_OPTIONS = 'POSITION_FETCHER_OPTIONS';
 export type PositionOptions = {
   includeInTvl?: boolean;
 };
-
-export const buildAppPositionsCacheKey = (opts: {
-  type: ContractType;
-  network: Network;
-  appId: string;
-  groupId: string;
-}) => `apps-v3:${opts.type}:${opts.network}:${opts.appId}:${opts.groupId}`;
 
 export const PositionFetcher =
   (type: ContractType) =>
@@ -43,12 +33,6 @@ export const PositionFetcher =
       SetMetadata(POSITION_FETCHER_NETWORK, network),
       SetMetadata(POSITION_FETCHER_TYPE, type),
       SetMetadata(POSITION_FETCHER_OPTIONS, options),
-      CacheOnIntervalBuilder<IPositionFetcher<Position>>({
-        targetMethod: 'getPositions',
-        key: buildAppPositionsCacheKey({ type, network, appId, groupId }),
-        timeout: process.env.NODE_ENV === 'production' ? 45 * 1000 : 45 * 1000,
-        failOnMissingData: process.env.NODE_ENV === 'production',
-      }),
       Injectable,
     );
   };
