+++ src/position/position.module.ts
@@ -1,6 +1,8 @@
 import { Module } from '@nestjs/common';
 import { DiscoveryModule } from '@nestjs/core';
 
+import { CacheModule } from '~cache/cache.module';
+
 import { PositionBalanceFetcherRegistry } from './position-balance-fetcher.registry';
 import { PositionFetcherRegistry } from './position-fetcher.registry';
 import { PositionSources } from './position-source';
@@ -8,7 +10,7 @@ import { PositionController } from './position.controller';
 import { PositionService } from './position.service';
 
 @Module({
-  imports: [DiscoveryModule],
+  imports: [DiscoveryModule, CacheModule],
   providers: [...PositionSources, PositionService, PositionFetcherRegistry, PositionBalanceFetcherRegistry],
   controllers: [PositionController],
   exports: [PositionService, PositionFetcherRegistry, PositionBalanceFetcherRegistry],
